Restore relationship report multiple custom data option. #27
In 4.6 the crmselect select field library was removed. This was being used to add a sortable select option.

This commit fixes the functionality to custom fields from more that one contact can be used, but the UI doesn't make much sense
at the moment as it is on a different tab but in only one field, which does not have the rich functionality that drove earlier decisions.

Future work might bring back sortability of output fields & either through tabs or other UI features better sort the available fields

m