Fix rollup to show total for second row
This might yet turn out to be a whack-a-mole. I was seeing a group-by on 2 different
fields and the first subtotal not showing as a subtotal when there was only one variation on the first group by.

But, the thing I removed might have helped in another case ... I kinda need to think of a better method