Discontinue support for 4.4 & 4.6
The change is administrative at this stage. The intention is that support for those versions ends at the last released version of this extension. I will remove some of the supporting legacy code over time but for now they should still work on the latest but I am signalling my intent to no longer support them.

Note that until such time as a version specific version of extension support comes out the supported version is defined as the latest point release of 4.7