added non deductible amount
I'm not sure how else to get this non deductible amount out of a
CiviCRM database. If you set a price set with a certain amount
that should be non-deductible, the line item table seems to be the
only place it is stored, yet it is not displayed any where else.