Fix alterCustomDataDisplay() to respect hook_civicrm_customFieldOptions()
Building the $fieldValueMap directly from the database doesn't respect custom field options defined via the core hook.  Based on the fix for #273, I updated this section similarly.

I suspect this whole section could be cleaned up a little more to use the API, but I wanted to keep this PR as simple as possible.