Fix tab weight not being set for reportlets
This fixes an issue where the tab weight for reportlets is not set
in the menu hook, causing weird tab sorting in some scenarios where
other extensions use the same hook.

The code already attempts to set the weight, but does so in the URL
array rather than the outer menu array.