enable search based on whether it's a primary or inherited membership
This filter is exposed in the UI, but using it has no effect because
the filter is not properly picked up.

This PR is a bit hacky, but I just couldn't figure out a more elegant
way to handle it since the search itself doesn't fit any defined
patterns in CiviReport.