ensure tag filter works on relationship report
In the core function whereTagClause, the $entity_table value
comes from $this->_tagFilterTable.

In the RelationshipExtended report, that variable is set to:
contact_a_civicrm_contact.

This variable ends up doing double duty in this functioon.

It provides both the field name to filter on (should be
contact_a_civicrm_contact - which gets set correctly).

In addition, it also determines which value should be filtered against
in the civicrm_entity_tag table (should be civicrm_contact - woops
it's getting contact_a_civicrm_contact instead!).

In this miserable hack I'm overriding the entire function just to
hard-code "WHERE entity_table = 'civicrm_contact'". I'm sure there is a
more elegant solution, but am not sure what it is exactly. Suggestions?