Set Main Page and Documentation URLs in info.xml
Hi Eileen, setting these URLs (well, at least the Main Extension Page) is helpful to me when I inspect an instance’s extensions list and want to know what a new release has to offer, and perhaps decide to update to a version that is not the most recent one. Since I use this extension on a few instances, here is a PR. I didn’t go through all your extensions, this issue might exist in others.